+++ test/magic_parameters_test.cpp
@@ -11,8 +11,8 @@ using namespace recognition;
 TEST(magic_parameters_test, closed_magic_set_parameter_get_parameter)
 {
     magic m;
-    EXPECT_THROW(m.set_parameter(magic::Parameter::BytesMax, 1), magic_is_closed);
-    EXPECT_THROW([[maybe_unused]] auto _ = m.get_parameter(magic::Parameter::BytesMax), magic_is_closed);
+    EXPECT_THROW(m.set_parameter(magic::parameters::BytesMax, 1), magic_is_closed);
+    EXPECT_THROW([[maybe_unused]] auto _ = m.get_parameter(magic::parameters::BytesMax), magic_is_closed);
     EXPECT_THROW([[maybe_unused]] auto _ = m.get_parameters(), magic_is_closed);
 }
 
@@ -22,16 +22,16 @@ TEST(magic_parameters_test, opened_magic_set_parameter_get_parameter)
     m.open(magic::Flag::Mime);
     std::mt19937 eng{std::random_device{}()};
     std::uniform_int_distribution<std::size_t> dist{0, 100};
-    magic::Parameters expected_parameters{
-        {magic::Parameter::IndirMax, dist(eng)},
-        {magic::Parameter::NameMax, dist(eng)},
-        {magic::Parameter::ElfPhnumMax, dist(eng)},
-        {magic::Parameter::ElfShnumMax, dist(eng)},
-        {magic::Parameter::ElfNotesMax, dist(eng)},
-        {magic::Parameter::RegexMax, dist(eng)},
-        {magic::Parameter::BytesMax, dist(eng)},
-        {magic::Parameter::EncodingMax, dist(eng)},
-        {magic::Parameter::ElfShsizeMax, dist(eng)}
+    magic::parameter_value_map_t expected_parameters{
+        {magic::parameters::IndirMax, dist(eng)},
+        {magic::parameters::NameMax, dist(eng)},
+        {magic::parameters::ElfPhnumMax, dist(eng)},
+        {magic::parameters::ElfShnumMax, dist(eng)},
+        {magic::parameters::ElfNotesMax, dist(eng)},
+        {magic::parameters::RegexMax, dist(eng)},
+        {magic::parameters::BytesMax, dist(eng)},
+        {magic::parameters::EncodingMax, dist(eng)},
+        {magic::parameters::ElfShsizeMax, dist(eng)}
     };
     for (const auto& parameter_pair : expected_parameters){
         m.set_parameter(parameter_pair.first, parameter_pair.second);
@@ -43,15 +43,15 @@ TEST(magic_parameters_test, opened_magic_set_parameter_get_parameter)
 TEST(magic_parameters_test, parameter_to_string_conversion)
 {
     std::vector magic_parameters_with_names{
-        std::make_pair(magic::Parameter::IndirMax, "IndirMax"),
-        std::make_pair(magic::Parameter::NameMax, "NameMax"),
-        std::make_pair(magic::Parameter::ElfPhnumMax, "ElfPhnumMax"),
-        std::make_pair(magic::Parameter::ElfShnumMax, "ElfShnumMax"),
-        std::make_pair(magic::Parameter::ElfNotesMax, "ElfNotesMax"),
-        std::make_pair(magic::Parameter::RegexMax, "RegexMax"),
-        std::make_pair(magic::Parameter::BytesMax, "BytesMax"),
-        std::make_pair(magic::Parameter::EncodingMax, "EncodingMax"),
-        std::make_pair(magic::Parameter::ElfShsizeMax, "ElfShsizeMax")
+        std::make_pair(magic::parameters::IndirMax, "IndirMax"),
+        std::make_pair(magic::parameters::NameMax, "NameMax"),
+        std::make_pair(magic::parameters::ElfPhnumMax, "ElfPhnumMax"),
+        std::make_pair(magic::parameters::ElfShnumMax, "ElfShnumMax"),
+        std::make_pair(magic::parameters::ElfNotesMax, "ElfNotesMax"),
+        std::make_pair(magic::parameters::RegexMax, "RegexMax"),
+        std::make_pair(magic::parameters::BytesMax, "BytesMax"),
+        std::make_pair(magic::parameters::EncodingMax, "EncodingMax"),
+        std::make_pair(magic::parameters::ElfShsizeMax, "ElfShsizeMax")
     };
     for (const auto& parameter_pair : magic_parameters_with_names){
         EXPECT_EQ(to_string(parameter_pair.first), parameter_pair.second);
@@ -60,16 +60,16 @@ TEST(magic_parameters_test, parameter_to_string_conversion)
 
 TEST(magic_parameters_test, parameters_to_string_conversion)
 {
-    EXPECT_EQ(to_string(magic::Parameters{
-        {magic::Parameter::IndirMax, 1},
-        {magic::Parameter::NameMax, 2},
-        {magic::Parameter::ElfPhnumMax, 3},
-        {magic::Parameter::ElfShnumMax, 4},
-        {magic::Parameter::ElfNotesMax, 5},
-        {magic::Parameter::RegexMax, 6},
-        {magic::Parameter::BytesMax, 7},
-        {magic::Parameter::EncodingMax, 8},
-        {magic::Parameter::ElfShsizeMax, 9}
+    EXPECT_EQ(to_string(magic::parameter_value_map_t{
+        {magic::parameters::IndirMax, 1},
+        {magic::parameters::NameMax, 2},
+        {magic::parameters::ElfPhnumMax, 3},
+        {magic::parameters::ElfShnumMax, 4},
+        {magic::parameters::ElfNotesMax, 5},
+        {magic::parameters::RegexMax, 6},
+        {magic::parameters::BytesMax, 7},
+        {magic::parameters::EncodingMax, 8},
+        {magic::parameters::ElfShsizeMax, 9}
     }),
         "IndirMax: 1, "
         "NameMax: 2, "